refactor(macos): split SessionListView.swift into per-concern view files
CodeScene flagged SessionListView.swift (2233 LOC, 103 revisions) as a
hotspot. Inspection showed the size and churn came from several
independently-changing, minimally-coupled SwiftUI types bundled into one
file by history rather than by design — each historical feature (ETA,
quota chips, backchannel, cache-regression badge, background-agent badge,
task summaries) landed its edits in a different section of this file,
inflating both LOC and revision count without any of those sections
depending on each other's private state.

Move-only split, no logic changes:
- SessionRowView.swift: ContextBar, SessionRowView, FlowLayout,
  TaskListView, SessionControlButton, SessionActionButtons
- GroupView.swift: the recursive project/rig group renderer
- SubagentRowView.swift: SubagentRowView + the shared DurationClock/
  LiveDurationText (#690)
- CostTimeframe.swift: the day/week/month/year enum shared by both
  SessionListView and GroupView

Left in place: the tooltip AppKit bridge, DisplayMode, and the
SessionListView struct itself (header + quota-chip subsystem) — the
part actually touched by #788's History tab-switcher work — to keep
this change out of the way of that area's ongoing churn.

SessionListView.swift: 2233 -> 1191 LOC.
